Split comment.go
