+++ private_isu/webapp/golang/app.go
@@ -24,7 +24,6 @@ import (
 	_ "github.com/go-sql-driver/mysql"
 	"github.com/gorilla/sessions"
 	"github.com/izumin5210/ro"
-	"github.com/izumin5210/ro/types"
 	"github.com/jmoiron/sqlx"
 	"github.com/zenazn/goji"
 	"github.com/zenazn/goji/web"
@@ -71,32 +70,6 @@ type Post struct {
 	Comments     []*Comment
 	User         *User
 	CSRFToken    string
-}
-
-type Comment struct {
-	ro.Model
-	ID        int       `db:"id" redis:"id"`
-	PostID    int       `db:"post_id" redis:"post_id"`
-	UserID    int       `db:"user_id" redis:"user_id"`
-	Comment   string    `db:"comment" redis:"comment"`
-	CreatedAt time.Time `db:"created_at" redis:"created_at"`
-	User      *User
-}
-
-func (c *Comment) GetKeySuffix() string {
-	return fmt.Sprint(c.ID)
-}
-
-var CommentScorerMap = map[string]types.ScorerFunc{
-	"post": func(m types.Model) interface{} {
-		return m.(*Comment).PostID
-	},
-	"user": func(m types.Model) interface{} {
-		return m.(*Comment).UserID
-	},
-	"created_at": func(m types.Model) interface{} {
-		return m.(*Comment).CreatedAt.UnixNano()
-	},
 }
 
 func init() {
